+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [v1.63.0] - 2026-01-06
 
 - Update `PROTOVALIDATE` lint rule to support field mask rules.
 - Add LSP completion for field numbers.
@@ -1464,7 +1464,7 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.62.1...HEAD
+[v1.63.0]: https://github.com/bufbuild/buf/compare/v1.62.1...v1.63.0
 [v1.62.1]: https://github.com/bufbuild/buf/compare/v1.62.0...v1.62.1
 [v1.62.0]: https://github.com/bufbuild/buf/compare/v1.61.0...v1.62.0
 [v1.61.0]: https://github.com/bufbuild/buf/compare/v1.60.0...v1.61.0